+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/apis/ledger/service/TargetAmountSaveService.java
@@ -1,4 +1,4 @@
-package kr.co.pennyway.api.apis.ledge.service;
+package kr.co.pennyway.api.apis.ledger.service;
 
 import kr.co.pennyway.domain.domains.target.domain.TargetAmount;
 import kr.co.pennyway.domain.domains.target.service.TargetAmountService;
