+++ pennyway-app-external-api/src/test/java/kr/co/pennyway/api/apis/ledger/controller/TargetAmountControllerUnitTest.java
@@ -1,7 +1,7 @@
-package kr.co.pennyway.api.apis.ledge.controller;
+package kr.co.pennyway.api.apis.ledger.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import kr.co.pennyway.api.apis.ledge.usecase.TargetAmountUseCase;
+import kr.co.pennyway.api.apis.ledger.usecase.TargetAmountUseCase;
 import kr.co.pennyway.api.config.supporter.WithSecurityMockUser;
 import kr.co.pennyway.domain.domains.target.exception.TargetAmountErrorCode;
 import org.junit.jupiter.api.*;
